requirements.txt
asgiref==3.5.0
autopep8==1.6.0
black==22.1.0
click==8.0.3
colorama==0.4.4
Django==4.0.2
django-filter==21.1
django-rest-passwordreset==1.2.1
djangorestframework==3.13.1
flake8==4.0.1
importlib-metadata==4.11.1
Markdown==3.3.6
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.0
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2021.3
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.0.1
tzdata==2021.5
zipp==3.7.0
